requirements.txt
fastapi
uvicorn
pydantic
httpx
pyyaml
numpy
scipy
pandas
